requirements.txt
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
decorator==5.1.1
dnspython==2.3.0
email-validate==1.1.2
et-xmlfile==1.1.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.45.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
grpcio==1.54.2
httplib2==0.21.0
idna==2.10
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
paramiko==2.12.0
prettytable==3.8.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfiglet==0.8.post1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
tqdm==4.64.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
yandex==0.0.2
yandex-cloud-client==1.0.5
yandexcloud==0.213.0
asgiref==3.7.2
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
decorator==5.1.1
Django==4.2.4
djangorestframework==3.14.0
dnspython==2.3.0
email-validate==1.1.2
et-xmlfile==1.1.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.45.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
grpcio==1.54.2
httplib2==0.21.0
idna==2.10
numpy==1.25.1
openpyxl==3.1.2
pandas==2.0.3
paramiko==2.12.0
prettytable==3.8.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfiglet==0.8.post1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.64.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.6
yandex==0.0.2
yandex-cloud-client==1.0.5
yandexcloud==0.213.0
